update add book

+++ controllers/book.js
@@ -1,4 +1,5 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 
 exports.getAllBooks = (req, res, next) => {
@@ -13,15 +14,31 @@ exports.getOneBook = (req, res, next) => {
         .catch(error => res.status(404).json({ error }));
     };
 
+exports.getBestRating = (req, res, next) => {
+    Book.find({ averageRating: { $gte: 3 } })
+        .then(books => res.status(200).json(books))
+        .catch(error => res.status(400).json({ error }));
+    }
+
+exports.createRating = (req, res, next) => {
+    Book.updateOne({ _id: req.params.id }, { $push: { ratings: req.body } })
+        .then(() => res.status(200).json({ message: 'Objet modifié !'}))
+        .catch(error => res.status(400).json({ error }));
+    }
+
 exports.createBook = (req, res, next) => {
-    delete req.body._id;
+    const bookObject = JSON.parse(req.body.book);
+    delete bookObject._id;
+    delete bookObject._userId;
     const book = new Book({
-        ...req.body
+        ...bookObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
     book.save()
         .then(() => res.status(201).json({ message: 'Objet enregistré !'}))
         .catch(error => res.status(400).json({ error }));
-    };
+    }
 
 exports.updateBook = (req, res, next) => {
     Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
@@ -29,8 +46,40 @@ exports.updateBook = (req, res, next) => {
         .catch(error => res.status(400).json({ error }));
     };
 
+exports.modifyBook = (req, res, next) => {
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+
+    delete bookObject._userId;
+    Book.findOne({ _id: req.params.id })
+        .then(book => {
+            if (book.userId !== req.auth.userId) {
+                res.status(401).json({ message: 'Non-autorisé !'});
+            } else {
+                Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'Objet modifié !'}))
+                .catch(error => res.status(401).json({ error }));
+            }
+        });
+};
+
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-        .catch(error => res.status(400).json({ error }));
-    };
+    Book.findOne({ _id: req.params.id })
+        .then(book => {
+            if (book.userId !== req.auth.userId) {
+                res.status(401).json({ message: 'Non-autorisé !' });
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
+                        .catch(error => res.status(400).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
+};
